refactor(wallet): create state module
